move builder to separate file
