+++ src/daft-distributed/src/pipeline_node/limit.rs
@@ -20,6 +20,7 @@ pub(crate) struct LimitNode {
 }
 
 impl LimitNode {
+    #[allow(dead_code)]
     pub fn new(
         limit: usize,
         local_physical_plans: Vec<LocalPhysicalPlanRef>,
@@ -41,6 +42,7 @@ impl LimitNode {
         }
     }
 
+    #[allow(dead_code)]
     async fn execution_loop(
         _task_dispatcher_handle: TaskDispatcherHandle,
         _local_physical_plans: Vec<LocalPhysicalPlanRef>,
